pyproject.toml
[tool.poetry]
name = "mkii-rl"
version = "0.1.0"
description = ""
authors = ["Dheemanth Manur <dmanu006@ucr.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8.0,<3.11"
stable-retro = "^0.9.2"
ipykernel = "^6.29.4"
stable-baselines3 = "^2.3.2"
opencv-python = "^4.10.0.82"
tensorboard = "2.13"
moviepy = "^1.0.3"
swig = "^4.2.1"
tqdm = "^4.66.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
